refactor: restructure wallet files
