+++ packages/graz/src/actions/wallet/wallet-connect/index.ts
@@ -1,4 +1,4 @@
-import type { AminoSignResponse, StdSignature } from "@cosmjs/amino";
+import type { AminoSignResponse } from "@cosmjs/amino";
 import { fromBech32 } from "@cosmjs/encoding";
 import type { AccountData, Algo, DirectSignResponse } from "@cosmjs/proto-signing";
 import type { Keplr, Key } from "@keplr-wallet/types";
@@ -8,269 +8,12 @@ import { getSdkError } from "@walletconnect/utils";
 // eslint-disable-next-line import/no-named-as-default
 import Long from "long";
 
-import { RECONNECT_SESSION_KEY } from "../constant";
-import { grazSessionDefaultValues, useGrazInternalStore, useGrazSessionStore } from "../store";
-import type { Wallet } from "../types/wallet";
-import { WALLET_TYPES, WalletType } from "../types/wallet";
-import { isAndroid, isIos, isMobile } from "../utils/os";
-import { promiseWithTimeout } from "../utils/timeout";
-
-/**
- * Function to check whether given {@link WalletType} or default configured wallet exists.
- *
- * @example
- * ```ts
- * const isSupported = checkWallet();
- * const isKeplrSupported = checkWallet("keplr");
- * ```
- */
-export const checkWallet = (type: WalletType = useGrazInternalStore.getState().walletType): boolean => {
-  try {
-    getWallet(type);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
-
-const clearSession = () => {
-  window.sessionStorage.removeItem(RECONNECT_SESSION_KEY);
-  useGrazSessionStore.setState(grazSessionDefaultValues);
-};
-
-/**
- * Function to return {@link Wallet} object and throws and error if it does not exist on `window`.
- *
- * @example
- * ```ts
- * try {
- *   const keplr = getKeplr();
- * } catch (error: Error) {
- *   console.error(error.message);
- * }
- * ```
- *
- * @see https://docs.keplr.app
- */
-export const getKeplr = (): Wallet => {
-  if (typeof window.keplr !== "undefined") {
-    const keplr = window.keplr;
-    const subscription: (reconnect: () => void) => void = (reconnect) => {
-      window.addEventListener("keplr_keystorechange", () => {
-        clearSession();
-        reconnect();
-      });
-      return () => {
-        window.removeEventListener("keplr_keystorechange", () => {
-          clearSession();
-          reconnect();
-        });
-      };
-    };
-    const res = Object.assign(keplr, {
-      subscription,
-    });
-    return res;
-  }
-
-  useGrazInternalStore.getState()._notFoundFn();
-  throw new Error("window.keplr is not defined");
-};
-
-/**
- * Function to return Leap object (which is {@link Wallet}) and throws and error if it does not exist on `window`.
- *
- * @example
- * ```ts
- * try {
- *   const leap = getLeap();
- * } catch (error: Error) {
- *   console.error(error.message);
- * }
- * ```
- *
- * @see https://docs.leapwallet.io/cosmos/for-dapps-connect-to-leap/add-leap-to-existing-keplr-integration
- */
-export const getLeap = (): Wallet => {
-  if (typeof window.leap !== "undefined") {
-    const leap = window.leap;
-    const subscription: (reconnect: () => void) => void = (reconnect) => {
-      window.addEventListener("leap_keystorechange", () => {
-        clearSession();
-        reconnect();
-      });
-      return () => {
-        window.removeEventListener("leap_keystorechange", () => {
-          clearSession();
-          reconnect();
-        });
-      };
-    };
-    const res = Object.assign(leap, {
-      subscription,
-    });
-    return res;
-  }
-
-  useGrazInternalStore.getState()._notFoundFn();
-  throw new Error("window.leap is not defined");
-};
-
-/**
- * Function to return cosmostation object (which is {@link Wallet}) and throws and error if it does not exist on `window`.
- *
- * @example
- * ```ts
- * try {
- *   const cosmostation = getCosmostation();
- * } catch (error: Error) {
- *   console.error(error.message);
- * }
- * ```
- *
- * @see https://docs.cosmostation.io/integration-extension/cosmos/integrate-keplr
- */
-export const getCosmostation = (): Wallet => {
-  if (typeof window.cosmostation.providers.keplr !== "undefined") {
-    const cosmostation = window.cosmostation.providers.keplr;
-    const subscription: (reconnect: () => void) => void = (reconnect) => {
-      window.cosmostation.cosmos.on("accountChanged", () => {
-        clearSession();
-        reconnect();
-      });
-      return () => {
-        window.cosmostation.cosmos.off("accountChanged", () => {
-          clearSession();
-          reconnect();
-        });
-      };
-    };
-    const res = Object.assign(cosmostation, {
-      subscription,
-    });
-    return res;
-  }
-
-  useGrazInternalStore.getState()._notFoundFn();
-  throw new Error("window.cosmostation.providers.keplr is not defined");
-};
-
-/**
- * Function to return {@link Wallet} object and throws and error if it does not exist on `window`.
- *
- * @example
- * ```ts
- * try {
- *   const vectis = getVectis();
- * } catch (error: Error) {
- *   console.error(error.message);
- * }
- * ```
- *
- *
- */
-export const getVectis = (): Wallet => {
-  if (typeof window.vectis !== "undefined") {
-    const vectis = window.vectis.cosmos;
-    const subscription: (reconnect: () => void) => void = (reconnect) => {
-      window.addEventListener("vectis_accountChanged", () => {
-        clearSession();
-        reconnect();
-      });
-      return () => {
-        window.removeEventListener("vectis_accountChanged", () => {
-          clearSession();
-          reconnect();
-        });
-      };
-    };
-    const getOfflineSignerOnlyAmino = (...args: Parameters<Wallet["getOfflineSignerOnlyAmino"]>) => {
-      return vectis.getOfflineSignerAmino(...args);
-    };
-
-    const experimentalSuggestChain = async (...args: Parameters<Wallet["experimentalSuggestChain"]>) => {
-      const [chainInfo] = args;
-      const adaptChainInfo = {
-        ...chainInfo,
-        rpcUrl: chainInfo.rpc,
-        restUrl: chainInfo.rest,
-        prettyName: chainInfo.chainName.replace(" ", ""),
-        bech32Prefix: chainInfo.bech32Config.bech32PrefixAccAddr,
-      };
-      return vectis.suggestChains([adaptChainInfo]);
-    };
-
-    const getKey = async (chainId: string): Promise<Key> => {
-      const key = await vectis.getKey(chainId);
-      return {
-        address: fromBech32(key.address).data,
-        algo: key.algo,
-        bech32Address: key.address,
-        name: key.name,
-        pubKey: key.pubKey,
-        isKeystone: false,
-        isNanoLedger: key.isNanoLedger,
-      };
-    };
-
-    const signDirect = async (...args: SignDirectParams): Promise<DirectSignResponse> => {
-      const { 1: signer, 2: signDoc } = args;
-      return vectis.signDirect(signer, {
-        bodyBytes: signDoc.bodyBytes || Uint8Array.from([]),
-        authInfoBytes: signDoc.authInfoBytes || Uint8Array.from([]),
-        accountNumber: Long.fromString(signDoc.accountNumber?.toString() || "", false),
-        chainId: signDoc.chainId || "",
-      });
-    };
-
-    const signAmino = async (...args: SignAminoParams): Promise<AminoSignResponse> => {
-      const { 1: signer, 2: signDoc } = args;
-      return vectis.signAmino(signer, signDoc);
-    };
-
-    return {
-      enable: (chainId: string) => vectis.enable(chainId),
-      getOfflineSigner: (chainId: string) => vectis.getOfflineSigner(chainId),
-      getOfflineSignerAuto: (chainId: string) => vectis.getOfflineSignerAuto(chainId),
-      getKey,
-      subscription,
-      getOfflineSignerOnlyAmino,
-      experimentalSuggestChain,
-      signDirect,
-      signAmino,
-    };
-  }
-
-  useGrazInternalStore.getState()._notFoundFn();
-  throw new Error("window.vectis is not defined");
-};
-
-type SignDirectParams = Parameters<Wallet["signDirect"]>;
-type SignAminoParams = Parameters<Wallet["signAmino"]>;
-
-interface WalletConnectSignDirectSigned {
-  chainId: string;
-  accountNumber: string;
-  authInfoBytes: string;
-  bodyBytes: string;
-}
-
-interface WalletConnectSignDirectResponse {
-  signature: StdSignature;
-  signed: WalletConnectSignDirectSigned;
-}
-
-export interface GetWalletConnectParams {
-  encoding: BufferEncoding;
-  walletType: WalletType.WC_KEPLR_MOBILE | WalletType.WC_LEAP_MOBILE;
-  appUrl: {
-    mobile: {
-      ios: string;
-      android: string;
-    };
-  };
-  formatNativeUrl: (appUrl: string, wcUri: string, os?: "android" | "ios") => string;
-}
+import { useGrazInternalStore, useGrazSessionStore } from "../../../store";
+import type { SignAminoParams, SignDirectParams, Wallet } from "../../../types/wallet";
+import { isAndroid, isIos, isMobile } from "../../../utils/os";
+import { promiseWithTimeout } from "../../../utils/timeout";
+import { clearSession } from "..";
+import type { GetWalletConnectParams, WalletConnectSignDirectResponse } from "./types";
 
 export const getWalletConnect = (params?: GetWalletConnectParams): Wallet => {
   if (!useGrazInternalStore.getState().walletConnect?.options?.projectId?.trim()) {
@@ -693,143 +436,4 @@ export const getWalletConnect = (params?: GetWalletConnectParams): Wallet => {
     subscription,
     init,
   };
-};
-
-export const getWCKeplr = (): Wallet => {
-  if (!useGrazInternalStore.getState().walletConnect?.options?.projectId?.trim()) {
-    throw new Error("walletConnect.options.projectId is not defined");
-  }
-
-  if (!isMobile()) throw new Error("WalletConnect Keplr mobile is only supported in mobile");
-
-  const params: GetWalletConnectParams = {
-    encoding: "base64",
-    appUrl: {
-      mobile: {
-        ios: "keplrwallet://",
-        android: "intent://",
-      },
-    },
-    walletType: WalletType.WC_KEPLR_MOBILE,
-    formatNativeUrl: (appUrl, wcUri, os) => {
-      const plainAppUrl = appUrl.replaceAll("/", "").replaceAll(":", "");
-      const encoded = encodeURIComponent(wcUri);
-      switch (os) {
-        case "ios":
-          return `${plainAppUrl}://wcV2?${encoded}`;
-        case "android":
-          return `${plainAppUrl}://wcV2?${encoded}#Intent;package=com.chainapsis.keplr;scheme=keplrwallet;end;`;
-        default:
-          return `${plainAppUrl}://wc?uri=${encoded}`;
-      }
-    },
-  };
-
-  return getWalletConnect(params);
-};
-
-export const getWCLeap = (): Wallet => {
-  if (!useGrazInternalStore.getState().walletConnect?.options?.projectId?.trim()) {
-    throw new Error("walletConnect.options.projectId is not defined");
-  }
-
-  if (!isMobile()) throw new Error("WalletConnect Leap mobile is only supported in mobile");
-
-  const params: GetWalletConnectParams = {
-    encoding: "base64",
-    appUrl: {
-      mobile: {
-        ios: "leapcosmos://",
-        android: "intent://",
-      },
-    },
-    walletType: WalletType.WC_LEAP_MOBILE,
-    formatNativeUrl: (appUrl, wcUri, os) => {
-      const plainAppUrl = appUrl.replaceAll("/", "").replaceAll(":", "");
-      const encoded = encodeURIComponent(wcUri);
-      switch (os) {
-        case "ios":
-          return `${plainAppUrl}://wcV2?${encoded}`;
-        case "android":
-          return `${plainAppUrl}://wcV2?${encoded}#Intent;package=io.leapwallet.cosmos;scheme=leapwallet;end;`;
-        default:
-          return `${plainAppUrl}://wc?uri=${encoded}`;
-      }
-    },
-  };
-
-  return getWalletConnect(params);
-};
-
-export const getWCCosmostation = (): Wallet => {
-  if (!useGrazInternalStore.getState().walletConnect?.options?.projectId?.trim()) {
-    throw new Error("walletConnect.options.projectId is not defined");
-  }
-
-  if (!isMobile()) throw new Error("WalletConnect Leap mobile is only supported in mobile");
-
-  const params: GetWalletConnectParams = {
-    encoding: "hex",
-    appUrl: {
-      mobile: {
-        ios: "cosmostation://",
-        android: "cosmostation://",
-      },
-    },
-    walletType: WalletType.WC_LEAP_MOBILE,
-    formatNativeUrl: (appUrl, wcUri, _os) => {
-      const plainAppUrl = appUrl.replaceAll("/", "").replaceAll(":", "");
-      return `${plainAppUrl}://wc?${wcUri}`;
-    },
-  };
-
-  return getWalletConnect(params);
-};
-
-/**
- * Function to return wallet object based on given {@link WalletType} or from store and throws an error if it does not
- * exist on `window` or unknown wallet type.
- *
- * @example
- * ```ts
- * const wallet = getWallet();
- * const keplr = getWallet("keplr");
- * ```
- *
- * @see {@link getKeplr}
- */
-export const getWallet = (type: WalletType = useGrazInternalStore.getState().walletType): Wallet => {
-  switch (type) {
-    case WalletType.KEPLR: {
-      return getKeplr();
-    }
-    case WalletType.LEAP: {
-      return getLeap();
-    }
-    case WalletType.COSMOSTATION: {
-      return getCosmostation();
-    }
-    case WalletType.VECTIS: {
-      return getVectis();
-    }
-    case WalletType.WALLETCONNECT: {
-      return getWalletConnect();
-    }
-    case WalletType.WC_KEPLR_MOBILE: {
-      return getWCKeplr();
-    }
-    case WalletType.WC_LEAP_MOBILE: {
-      return getWCLeap();
-    }
-    case WalletType.WC_COSMOSTATION_MOBILE: {
-      return getWCCosmostation();
-    }
-    default: {
-      throw new Error("Unknown wallet type");
-    }
-  }
-};
-
-export const getAvailableWallets = (): Record<WalletType, boolean> => {
-  return Object.fromEntries(WALLET_TYPES.map((type) => [type, checkWallet(type)])) as Record<WalletType, boolean>;
 };
