+++ packages/graz/src/index.ts
@@ -3,6 +3,14 @@ export * from "./actions/chains";
 export * from "./actions/configure";
 export * from "./actions/methods";
 export * from "./actions/wallet";
+export * from "./actions/wallet/cosmostation";
+export * from "./actions/wallet/keplr";
+export * from "./actions/wallet/leap";
+export * from "./actions/wallet/vectis";
+export * from "./actions/wallet/wallet-connect";
+export * from "./actions/wallet/wallet-connect/cosmostation";
+export * from "./actions/wallet/wallet-connect/keplr";
+export * from "./actions/wallet/wallet-connect/leap";
 export * from "./adapter";
 export * from "./chains";
 export * from "./hooks/account";
